Resolve circular dependency that emerged after some functions were moved.
